refactor: step1 피드백 반영-themeItems로 items 따로 저장해 상위 컴포넌트에서 데이터 관리
